requirements.txt
discord
python-dotenv
pillow
youtube_dl